add (very simple) tests around hook service enable/disable
